requirements.txt
torch>=1.1.0
torchvision>=0.3.0
visdom>=0.1.8.9
PyYAML>=6.0
scikit-learn>=0.23.2
imbalanced-learn>=0.8.1
dominate>=2.5.1
h5py>=3.4.0
matplotlib>=3.3.1
pandas>=1.3.3
SimpleITK>=2.1.1
albumentations>=1.0.3